Move remaining fragments into fragment package
